repo-updater: Move main() to a shared.Main() function
